requirements-prod.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
anyio==3.6.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.88.0
frozenlist==1.3.3
pydantic==1.10.2
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.45
uvicorn==0.20.0